+++ CHANGELOG.md
@@ -15,8 +15,10 @@ HEAD
   per page in services like `FindItem` that support paging.
 - Support creating a proper response when getting a notification request
   on the callback URL of a push subscription.
-- `FolderCollection.subscribe_to_[pull|push|streaming]` now return a single
+- `FolderCollection.subscribe_to_[pull|push|streaming]()` now return a single
   subscription instead of a 1-element generator.
+- `FolderCollection` now has the same `[pull|push|streaming]_subscription()`
+  context managers as folders.
 
 
 4.6.2